app/requirements.txt
Flask
mysql-connector
flask-restplus
werkzeug==0.16.1

